extrated naming from Relationship and Node into HasVariableTrait
