+++ src/Traits/RelationshipTrait.php
@@ -24,7 +24,9 @@ namespace WikibaseSolutions\CypherDSL\Traits;
 /**
  * This trait should be used by any expression that returns a relationship.
  */
-trait PathTypeTrait
+trait RelationshipTrait
 {
     use StructuralTypeTrait;
+    use HasPropertiesTrait;
+    use HasVariableTrait;
 }
